# src/Maestro/Client/generator/generated/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "2018-07-16",
    "title": "Version 2018-07-16"
  },
  "host": "localhost:4430",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/subscriptions": {
      "get": {
        "tags": [
          "Subscriptions"
        ],
        "operationId": "Subscriptions_GetAllSubscriptions",
        "consumes": [],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "sourceRepository",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "targetRepository",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "channelId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The api version",
            "required": true,
            "type": "string",
            "enum": [
              "2018-07-16"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": {
                "$ref": "#/definitions/Subscription"
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Subscriptions"
        ],
        "operationId": "Subscriptions_Create",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "subscription",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/SubscriptionData"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The api version",
            "required": true,
            "type": "string",
            "enum": [
              "2018-07-16"
            ]
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Subscription"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/api/subscriptions/{id}": {
      "get": {
        "tags": [
          "Subscriptions"
        ],
        "operationId": "Subscriptions_GetSubscription",
        "consumes": [],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The api version",
            "required": true,
            "type": "string",
            "enum": [
              "2018-07-16"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Subscription"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Subscriptions"
        ],
        "operationId": "Subscriptions_DeleteSubscription",
        "consumes": [],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The api version",
            "required": true,
            "type": "string",
            "enum": [
              "2018-07-16"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Subscription"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Subscriptions"
        ],
        "operationId": "Subscriptions_UpdateSubscription",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "update",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/SubscriptionUpdate"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The api version",
            "required": true,
            "type": "string",
            "enum": [
              "2018-07-16"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Subscription"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/api/subscriptions/{id}/history": {
      "get": {
        "tags": [
          "Subscriptions"
        ],
        "operationId": "Subscriptions_GetSubscriptionHistory",
        "consumes": [],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "api-version",
            "in": "query",
            "description": "The api version",
            "required": true,
            "type": "string",
            "enum": [
              "2018-07-16"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": {
                "$ref": "#/definitions/SubscriptionHistoryItem"
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Asset": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer",
          "readOnly": false
        },
        "name": {
          "type": "string",
          "readOnly": false
        },
        "version": {
          "type": "string",
          "readOnly": false
        },
        "locations": {
          "uniqueItems": false,
          "type": "array",
          "items": {
            "$ref": "#/definitions/AssetLocation"
          },
          "readOnly": false
        }
      }
    },
    "AssetLocation": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer",
          "readOnly": false
        },
        "location": {
          "type": "string",
          "readOnly": false
        },
        "type": {
          "enum": [
            "none",
            "nugetFeed",
            "container"
          ],
          "type": "string",
          "readOnly": false
        }
      }
    },
    "ApiError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "readOnly": false
        },
        "errors": {
          "uniqueItems": false,
          "type": "array",
          "items": {
            "type": "string"
          },
          "readOnly": false
        }
      }
    },
    "Build": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer",
          "readOnly": false
        },
        "repository": {
          "type": "string",
          "readOnly": false
        },
        "commit": {
          "type": "string",
          "readOnly": false
        },
        "buildNumber": {
          "type": "string",
          "readOnly": false
        },
        "dateProduced": {
          "format": "date-time",
          "type": "string",
          "readOnly": false
        },
        "channels": {
          "uniqueItems": false,
          "type": "array",
          "items": {
            "$ref": "#/definitions/Channel"
          },
          "readOnly": false
        },
        "assets": {
          "uniqueItems": false,
          "type": "array",
          "items": {
            "$ref": "#/definitions/Asset"
          },
          "readOnly": false
        },
        "dependencies": {
          "uniqueItems": false,
          "type": "array",
          "items": {
            "$ref": "#/definitions/BuildRef"
          },
          "readOnly": false
        }
      }
    },
    "Channel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer",
          "readOnly": false
        },
        "name": {
          "type": "string",
          "readOnly": false
        },
        "classification": {
          "type": "string",
          "readOnly": false
        }
      }
    },
    "BuildRef": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Subscription": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string",
          "readOnly": false
        },
        "channel": {
          "$ref": "#/definitions/Channel",
          "readOnly": false
        },
        "sourceRepository": {
          "type": "string",
          "readOnly": false
        },
        "targetRepository": {
          "type": "string",
          "readOnly": false
        },
        "targetBranch": {
          "type": "string",
          "readOnly": false
        },
        "policy": {
          "$ref": "#/definitions/SubscriptionPolicy",
          "readOnly": false
        },
        "lastAppliedBuild": {
          "$ref": "#/definitions/Build",
          "readOnly": false
        }
      }
    },
    "SubscriptionPolicy": {
      "required": [
        "updateFrequency",
        "mergePolicy"
      ],
      "type": "object",
      "properties": {
        "updateFrequency": {
          "enum": [
            "none",
            "everyDay",
            "everyBuild"
          ],
          "type": "string"
        },
        "mergePolicy": {
          "enum": [
            "none",
            "buildSucceeded",
            "unitTestPassed",
            "never"
          ],
          "type": "string"
        }
      }
    },
    "SubscriptionData": {
      "required": [
        "channelName",
        "sourceRepository",
        "targetRepository",
        "targetBranch",
        "policy"
      ],
      "type": "object",
      "properties": {
        "channelName": {
          "type": "string"
        },
        "sourceRepository": {
          "type": "string"
        },
        "targetRepository": {
          "type": "string"
        },
        "targetBranch": {
          "type": "string"
        },
        "policy": {
          "$ref": "#/definitions/SubscriptionPolicy"
        }
      }
    },
    "SubscriptionUpdate": {
      "type": "object",
      "properties": {
        "channelName": {
          "type": "string"
        },
        "sourceRepository": {
          "type": "string"
        },
        "policy": {
          "$ref": "#/definitions/SubscriptionPolicy"
        }
      }
    },
    "SubscriptionHistoryItem": {
      "type": "object",
      "properties": {
        "timestamp": {
          "format": "date-time",
          "type": "string",
          "readOnly": false
        },
        "errorMessage": {
          "type": "string",
          "readOnly": false
        },
        "success": {
          "type": "boolean",
          "readOnly": false
        },
        "subscriptionId": {
          "format": "uuid",
          "type": "string",
          "readOnly": false
        },
        "action": {
          "type": "string",
          "readOnly": false
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}